Functional Balance component w/ css module.
