+++ app/components/shared/Balance/index.js
@@ -1,0 +1,1 @@
+export { default as Balance } from "./Balance";
